AED/requirements.txt
numpy
matplotlib
sklearn
librosa
python-Levenshtein
tensorflow